backend/requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
google-genai
fastapi
uvicorn